feat: Moving brushes to Ryot and using it in compass
