+++ ryot/src/bevy_ryot/drawing/brushes/square.rs
@@ -1,5 +1,4 @@
-use crate::{Brush, BrushItem};
-use ryot::position::TilePosition;
+use crate::bevy_ryot::drawing::brushes::*;
 
 pub struct Square;
