+++ ryot_compass/src/lib.rs
@@ -1,5 +1,3 @@
-#![feature(fn_traits)]
-#![feature(unboxed_closures)]
 #![feature(type_alias_impl_trait)]
 
 use bevy::app::{App, AppExit, Plugin};
